requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_bcrypt
flask_login
pymysql
python-dotenv
